feat: Move Builders in target class
